Move document color
